// ts/types/stats.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Token = Literal | Number | Enum | Unknown | NestedStat;

export interface StatsByFile {
  [k: string]: Stat;
}
export interface Stat {
  files: string[];
  generated_name: string;
  tokens: Token[];
  implied_stats?: {
    [k: string]: number;
  };
}
export interface Literal {
  type: "literal";
  value: string;
}
/**
 * Matches a number of the form /[+-]?\d*\.?\d+/
 */
export interface Number {
  type: "number";
  index: number;
  stat: string;
  stat_value_handlers?: string[];
}
/**
 * Integer stat values mapped to rows of a dat file.
 */
export interface Enum {
  type: "enum";
  index: number;
  stat: string;
  /**
   * Reference to the entry in stat_value_handlers.json where the enum values can be found.
   */
  stat_value_handler: string;
}
/**
 * How to handle this item is not known. At the current time only the 'sells for an additional unique <base item type>' crucible mod has an unknown token.
 */
export interface Unknown {
  type: "unknown";
  index: number;
  stat: string;
  stat_value_handler?: string;
}
/**
 * Matches any other stat - used by 'While a ... is in your Presence, <stat>' mods
 */
export interface NestedStat {
  type: "nested";
  added_stat: string;
}